Tweak intro1 start and finale4 ending criteria (fix #1, fix #2)
Using a position based approach with tick offset for both now.  The same
cylinder as before for the ending, and touching the point when getting out of
bed for the first time for the start.  Moving around pre-control and touching
that point will give incorrect values (too long), the fix: don't move.  The
ending still might be a tick off randomly, but whatever.

Validated against Znernicus's PB run.